Group the code and the e2e tests
