+++ crates/e2e/tests/e2e/replace_order.rs
@@ -23,12 +23,6 @@ async fn local_node_try_replace_someone_else_order() {
     run_test(try_replace_someone_else_order_test).await;
 }
 
-#[tokio::test]
-#[ignore]
-async fn local_node_single_replace_order_test_with_custom_app_data() {
-    run_test(single_replace_order_test_with_custom_app_data).await;
-}
-
 async fn try_replace_someone_else_order_test(web3: Web3) {
     let mut onchain = OnchainComponents::deploy(web3.clone()).await;
 
@@ -231,137 +225,16 @@ async fn single_replace_order_test(web3: Web3) {
     );
     let order_id = services.create_order(&order).await.unwrap();
 
-    // Replace order
-    let new_order = OrderCreation {
-        sell_token: token_a.address(),
-        sell_amount: to_wei(3),
-        buy_token: token_b.address(),
-        buy_amount: to_wei(1),
-        valid_to: model::time::now_in_epoch_seconds() + 300,
-        kind: OrderKind::Sell,
-        partially_fillable: false,
-        app_data: OrderCreationAppData::Full {
-            full: format!(
-                r#"{{"version":"1.1.0","metadata":{{"replacedOrder":{{"uid":"{}"}}}}}}"#,
-                order_id
-            ),
-        },
-        ..Default::default()
-    }
-    .sign(
-        EcdsaSigningScheme::Eip712,
-        &onchain.contracts().domain_separator,
-        SecretKeyRef::from(&SecretKey::from_slice(trader.private_key()).unwrap()),
-    );
-    services.create_order(&new_order).await.unwrap();
-
-    // Check the previous order is cancelled
-    let old_order = services.get_order(&order_id).await.unwrap();
-    assert_eq!(old_order.metadata.status, OrderStatus::Cancelled);
-
-    // Drive solution
-    tracing::info!("Waiting for trade.");
-    let balance_before = token_a.balance_of(trader.address()).call().await.unwrap();
-    wait_for_condition(TIMEOUT, || async { services.solvable_orders().await == 1 })
-        .await
-        .unwrap();
-
-    wait_for_condition(TIMEOUT, || async { services.solvable_orders().await == 0 })
-        .await
-        .unwrap();
-
-    let balance_after = token_a.balance_of(trader.address()).call().await.unwrap();
-    assert_eq!(
-        balance_before.checked_sub(balance_after).unwrap(),
-        to_wei(3)
-    );
-}
-
-async fn single_replace_order_test_with_custom_app_data(web3: Web3) {
-    let mut onchain = OnchainComponents::deploy(web3.clone()).await;
-
-    let [solver] = onchain.make_solvers(to_wei(1)).await;
-    let [trader] = onchain.make_accounts(to_wei(1)).await;
-    let [token_a, token_b] = onchain
-        .deploy_tokens_with_weth_uni_v2_pools(to_wei(1_000), to_wei(1_000))
-        .await;
-
-    // Fund trader accounts
-    token_a.mint(trader.address(), to_wei(30)).await;
-
-    // Create and fund Uniswap pool
-    token_a.mint(solver.address(), to_wei(1000)).await;
-    token_b.mint(solver.address(), to_wei(1000)).await;
-    tx!(
-        solver.account(),
-        onchain
-            .contracts()
-            .uniswap_v2_factory
-            .create_pair(token_a.address(), token_b.address())
-    );
-    tx!(
-        solver.account(),
-        token_a.approve(
-            onchain.contracts().uniswap_v2_router.address(),
-            to_wei(1000)
-        )
-    );
-    tx!(
-        solver.account(),
-        token_b.approve(
-            onchain.contracts().uniswap_v2_router.address(),
-            to_wei(1000)
-        )
-    );
-    tx!(
-        solver.account(),
-        onchain.contracts().uniswap_v2_router.add_liquidity(
-            token_a.address(),
-            token_b.address(),
-            to_wei(1000),
-            to_wei(1000),
-            0_u64.into(),
-            0_u64.into(),
-            solver.address(),
-            U256::max_value(),
-        )
-    );
-
-    // Approve GPv2 for trading
-    tx!(
-        trader.account(),
-        token_a.approve(onchain.contracts().allowance, to_wei(15))
-    );
-
-    // Place Orders
-    let services = Services::new(onchain.contracts()).await;
-    services.start_protocol(solver).await;
-
-    let order = OrderCreation {
-        sell_token: token_a.address(),
-        sell_amount: to_wei(10),
-        buy_token: token_b.address(),
-        buy_amount: to_wei(5),
-        valid_to: model::time::now_in_epoch_seconds() + 300,
-        kind: OrderKind::Sell,
-        ..Default::default()
-    }
-    .sign(
-        EcdsaSigningScheme::Eip712,
-        &onchain.contracts().domain_separator,
-        SecretKeyRef::from(&SecretKey::from_slice(trader.private_key()).unwrap()),
-    );
-    let order_id = services.create_order(&order).await.unwrap();
     let app_data = format!(
         r#"{{
-                    "version":"1.1.0",
-                    "metadata":{{
-                        "replacedOrder":{{
-                            "uid":"{}"
-                        }},
-                        "customStuff": 20
-                      }}
-                    }}"#,
+              "version":"1.1.0",
+                  "metadata":{{
+                      "replacedOrder":{{
+                          "uid":"{}"
+                      }},
+                      "customStuff": 20
+                  }}
+              }}"#,
         order_id
     );
 
@@ -385,6 +258,27 @@ async fn single_replace_order_test_with_custom_app_data(web3: Web3) {
         SecretKeyRef::from(&SecretKey::from_slice(trader.private_key()).unwrap()),
     );
     let new_order_uid = services.create_order(&new_order).await.unwrap();
+
+    // Check the previous order is cancelled
+    let old_order = services.get_order(&order_id).await.unwrap();
+    assert_eq!(old_order.metadata.status, OrderStatus::Cancelled);
+
+    // Drive solution
+    tracing::info!("Waiting for trade.");
+    let balance_before = token_a.balance_of(trader.address()).call().await.unwrap();
+    wait_for_condition(TIMEOUT, || async { services.solvable_orders().await == 1 })
+        .await
+        .unwrap();
+
+    wait_for_condition(TIMEOUT, || async { services.solvable_orders().await == 0 })
+        .await
+        .unwrap();
+
+    let balance_after = token_a.balance_of(trader.address()).call().await.unwrap();
+    assert_eq!(
+        balance_before.checked_sub(balance_after).unwrap(),
+        to_wei(3)
+    );
 
     // Check the previous order is cancelled
     let new_order = services.get_order(&new_order_uid).await.unwrap();
